Fixed issues with scaling on device rotate
When the iPad was rotated to portrait view the submit button would
essentially disappear due to the way that I had some constraints set in
the storyboard.  This issue should be resolved and now if the device is
rotated to portrait view the form will shrink and the submit will be
next to the phone number field.