Added second page
Added a secondary view controller as a sort of hidden admin interface.
This second view loads a php script that pulls all of the entries from
the database and displays them in a table for easy reference.  Learning
how to create multi page apps, this is my first attempt.  Clicking on
the small groups logo will take you to the “admin page” and swiping
left on the admin page will take you back to the form.